+++ R/saitem.R
@@ -27,7 +27,7 @@ NULL
 read_sai <- function(jsai) {
 
     #  if (! .jcall(jsai, "Z", "isProcessed"))
-    #    stop("You must run '.jws_compute()' on your workspace.")
+    #    stop("You must run 'jws_compute()' on your workspace.")
 
     jdef <- .jcall(jsai, "Ljdplus/sa/base/api/SaDefinition;", "getDefinition")
 
@@ -36,7 +36,7 @@ read_sai <- function(jsai) {
     if (!is.jnull(jestimation)) {
         jrslt <- .jcall(
             obj = jestimation,
-            returnSig = "Ljdplus/toolkit/base/api/information/Explorable;",
+            returnSig = "Ljdplus/toolkit/base/api/information/GenericExplorable;",
             method = "getResults"
         )
     }
@@ -100,7 +100,7 @@ read_sai <- function(jsai) {
     if (is.jnull(jestimation)) {
         return(NULL)
     }
-    jrslt <- .jcall(jestimation, "Ljdplus/toolkit/base/api/information/Explorable;", "getResults")
+    jrslt <- .jcall(jestimation, "Ljdplus/toolkit/base/api/information/GenericExplorable;", "getResults")
     if (is.null(items)) {
         items <- rjd3toolkit::.proc_dictionary2(jrslt)
     }
@@ -118,15 +118,15 @@ read_sai <- function(jsai) {
     if (is.jnull(jestimation)) {
         return(NULL)
     }
-    jrslt <- .jcall(jestimation, "Ljdplus/toolkit/base/api/information/Explorable;", "getResults")
+    jrslt <- .jcall(jestimation, "Ljdplus/toolkit/base/api/information/GenericExplorable;", "getResults")
     res <- rjd3toolkit::.jd3_object(jrslt, result = TRUE)
     return(res)
 }
 
 
-#' @name .jsap_name
+#' @name jsap_name
 #' @export
-.jsai_name <- function(jsai) {
+sai_name <- function(jsai) {
     return(.jcall(jsai, "S", "getName"))
 }