+++ NEWS.md
@@ -25,6 +25,23 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * `.jws_load()` in favour of `.jws_open()`
 * `.jsap_sai_count` is now deprecated
+* Rename `.jsap_sai_count()` in `sap_sai_count()`
+* Rename `.jws_sap_count()` in `ws_sap_count()`
+* Rename `.jws_open()` in `jws_open()`
+* Rename `.jread_ws()` in `jread_ws()`
+* Rename `.jread_sap()` in `jread_sap()`
+* Rename `.jws_new()` in `jws_new()`
+* Rename `.jws_sap_new()` in `.jws_sap_new()`
+* Rename `.jws_make_copy()` in `jws_make_copy()`
+* Rename `.jsap_make_copy()` in `jsap_make_copy()`
+* Rename `.jws_compute()` in `jws_compute()`
+* Rename `.jsap_sai` in `jsap_sai`
+* Rename `.jws_sap` in `jws_sap`
+* Rename `.jsai_name` in `sai_name`
+* Rename `.jsap_name` in `sap_name`
+* Rename `.jsap_sai_names` in `sap_sai_names`
+* Rename `.jsap_refresh` in `jsap_refresh`
+* Rename `.jws_refresh` in `jws_refresh`
 
 
 ## [3.2.4] - 2025-02-04